requirements.txt
discord_components
wavelink
discord.py